Add test suite for basic coverage
This commit adds a Jest-based test suite for the project to cover injecting and removing snippets from the code, as well as testing
configuration options.

This commit introduces a change to how the configuration file is loaded. Prior to this commit, configuration file path was hard-coded.

This commit allows an optional config file path. This optional path is not used in the main entrypoint program or exposed as an end-user option, but it's introduced so we can test the configuration loader code.

This commit relies on the `temporal/samples-sdk` repository for its
tests.

remove unnecessary pieces